+++ src/components/BulkEditPane/BulkEditListResult/BulkEditManualUploadModal/BulkEditManualUploadModal.test.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { render, screen, fireEvent, waitFor } from '@testing-library/react';
+import { render, screen, fireEvent, waitFor } from '@folio/jest-config-stripes/testing-library/react';
 import BulkEditManualUploadModal from './BulkEditManualUploadModal';