+++ src/hooks/useErrorType.test.js
@@ -1,4 +1,4 @@
-import { renderHook, act } from '@testing-library/react-hooks';
+import { renderHook, act } from '@folio/jest-config-stripes/testing-library/react-hooks';
 import { useErrorType } from './useErrorType';
 import { ERROR_TYPES } from '../constants';