+++ src/hooks/usePublishCoordinator.test.js
@@ -3,7 +3,7 @@ import {
   QueryClientProvider,
 } from 'react-query';
 
-import { renderHook } from '@testing-library/react-hooks';
+import { renderHook } from '@folio/jest-config-stripes/testing-library/react-hooks';
 import { useOkapiKy } from '@folio/stripes/core';
 
 import {